success to show adminAppPage
